New Sub Controlers and routes Improvements

+++ controllers/subControllers.js
@@ -91,13 +91,71 @@ export const updateSubscription = async (req, res, next) => {
 export const deleteSubscription = async (req, res, next) => {
   try {
     const subId = req.params.id;
-    const deleteSubscription = await Subscription.findByIdAndDelete(subId);
+    const subscription = await Subscription.findById(subId);
     if (!deleteSubscription) {
       let error = new Error("Subscription Not Found");
       error.statusCode = 404;
       throw error;
     }
-    res.status(200).json({ success: true, data: deleteSubscription });
+    if (req.user._id.toString() !== subscription.user.toString()) {
+      let error = new Error(" You are not the owner of this account");
+      error.statusCode = 401;
+      throw error;
+    }
+    const deletedSubscription = await Subscription.findByIdAndDelete(subId);
+    res.status(200).json({ success: true, data: deletedSubscription });
+  } catch (error) {
+    console.log(error);
+    next(error);
+  }
+};
+
+export const upcomingRenewals = async (req, res, next) => {
+  try {
+    const limit = req.query.limit ?? "month";
+    const subscriptions = await Subscription.find();
+    const upcomingRenewals = subscriptions.filter((sub) => {
+      const nextRenewalDate = new Date(sub.renewalDate);
+      const today = new Date(Date.now());
+      if (limit === "week") {
+        return nextRenewalDate >= today && nextRenewalDate <= today + 7;
+      } else if (limit === "today") {
+        return nextRenewalDate === today; // expiring today.
+      } else {
+        return nextRenewalDate >= today && nextRenewalDate <= today + 30; // by default 30 days limit
+      }
+    });
+    res.status(200).json({ success: true, data: upcomingRenewals });
+  } catch (error) {
+    console.log(error);
+    next(error);
+  }
+};
+
+export const cancelSubscrition = async (req, res, next) => {
+  try {
+    const subId = req.params.id; // getting the cancel subscription id
+    const subscription = await Subscription.findById(subId);
+    if (!subscription) {
+      let error = new Error("Subscription Not Found");
+      error.statusCode = 404;
+      throw error;
+    }
+    if (req.user._id.toString() !== subscription.user.toString()) {
+      let error = new Error(" You are not the owner of this account");
+      error.statusCode = 401;
+      throw error;
+    }
+    const updateSUbscription = await Subscription.findByIdAndUpdate(
+      id,
+      {
+        status: "cancelled",
+      },
+      {
+        new: true,
+      }
+    );
+    res.status(200).json({ success: true, data: updateSUbscription });
   } catch (error) {
     console.log(error);
     next(error);
